+++ CHANGELOG
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Offline capable entries won't hide on the drawer when offline.
+- Consolidated online check function.
 
 ## [0.2.11] - 2022-07-08
 ### Added
