+++ src/pages/home.tsx
@@ -1,25 +1,10 @@
 import { Typography, Box, Link } from '@mui/material';
-import { useEffect, useState } from 'react';
 import { useNavigate } from "react-router-dom";
+import { PageProps } from '../interfaces';
 import { HomeMenu } from "./segments/menu";
 
-export default function Home() {
-	const [connectionState, setConnectionState] = useState(true);
+export default function Home({online}:PageProps) {
 	const navigate = useNavigate();
-	const MINUTE_MS = 15000;
-
-	// https://stackoverflow.com/a/65049865
-	useEffect(() => {
-		const interval = setInterval(() => {
-			if (navigator.onLine) {
-				setConnectionState(true);
-			} else {
-				setConnectionState(false);
-			}
-		}, MINUTE_MS);
-
-		return () => clearInterval(interval);
-	}, []);
 
 	return (
 		<Box sx={{ flexGrow: 1 }}>
@@ -29,7 +14,7 @@ export default function Home() {
 				<Link sx={{ cursor: 'pointer', textDecoration: 'none' }} onClick={() => navigate('/help')}>see the help page</Link>.
 			</Typography>
 
-			<HomeMenu onlineState={connectionState} />
+			<HomeMenu onlineState={online} />
 		</Box>
 	)
 }
